moved http 500 error function and fixed linting errors
