+++ src/helpers/respond.ts
@@ -43,3 +43,13 @@ export const generate405 = async (req: Request, res: Response): Promise<void> =>
 		data: []
 	});
 };
+
+export const generate500 = async (req: Request, res: Response, error: Error): Promise<void> => {
+	res.json({
+		code: 500,
+		status: 'Internal Server Error',
+		description: 'Something Went Wrong',
+		data: []
+	});
+	throw error;
+};
